+++ backend/src/baserow/contrib/database/fields/registries.py
@@ -36,8 +36,8 @@ class FieldType(MapAPIExceptionsInstanceMixin, APIUrlsInstanceMixin,
         field_type_registry.register(ExampleFieldType())
     """
 
-    can_sort_in_view = True
-    """Indicates whether is is possible to sort on a field in a view."""
+    can_order_by = True
+    """Indicates whether it is possible to order by this field type."""
 
     def prepare_value_for_db(self, instance, value):
         """